# swagger.yaml
openapi: 3.0.3
info:
  title: Succulent Identifier API
  description: |
    REST API for the Succulent Identifier application. This API provides endpoints for:
    - Plant identification using machine learning
    - AI-powered chat assistance for plant care questions
    - Identification history management
    - Image serving

    Built with Go, PostgreSQL, OpenAI GPT-4o-mini, and PyTorch.
  version: 1.0.0
  contact:
    name: Succulent Identifier
    url: https://github.com/yourusername/succulent_identifier
servers:
  - url: http://localhost:8080
    description: Local development server (Backend API)
  - url: http://localhost:8000
    description: Local ML service
...
paths:
...
  /history:
    get:
      tags:
        - History
      summary: Get paginated list of identifications
      description: Retrieve a paginated list of past plant identifications
      operationId: getHistory
      parameters:
        - name: limit
          in: query
          description: Number of items to return (max 100)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Successful response with identification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
              example:
                items:
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    genus: "Haworthia"
                    species: "haworthia_zebrina"
                    confidence: 0.9468
                    image_path: "4cb08722-461a-4d6f-acd4-b06516cde3e8.jpg"
                    created_at: "2026-02-17T22:08:59Z"
                  - id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    genus: "Opuntia"
                    species: "opuntia_microdasys"
                    confidence: 0.9730
                    image_path: "5dc09833-572b-5e8f-bde5-g23456789012.jpg"
                    created_at: "2026-02-17T21:45:30Z"
                total: 4
                limit: 20
                offset: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /history/{id}:
    get:
      tags:
        - History
      summary: Get identification details
      description: Retrieve detailed information about a specific identification
      operationId: getHistoryById
      parameters:
        - name: id
          in: path
          description: Identification ID (UUID)
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with identification details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDetailResponse'
              example:
                id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                genus: "Haworthia"
                species: "haworthia_zebrina"
                confidence: 0.9468
                image_path: "4cb08722-461a-4d6f-acd4-b06516cde3e8.jpg"
                care_guide:
                  sunlight: "Bright, indirect light. Avoid direct sun."
                  watering: "Water when soil is completely dry (every 2-3 weeks)."
                  soil: "Well-draining cactus or succulent mix."
                  notes: "Hardy and easy to care for. Great for beginners."
                created_at: "2026-02-17T22:08:59Z"
        '404':
          description: Identification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - History
      summary: Soft delete an identification
      description: Soft deletes an identification by setting its deleted_at timestamp. The record is hidden but not permanently removed from the database.
      operationId: deleteHistoryById
      parameters:
        - name: id
          in: path
          description: Identification ID (UUID)
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful deletion
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Identification deleted successfully"
        '404':
          description: Identification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "Identification not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CareInstructions:
      type: object
      properties:
        sunlight:
          type: string
          description: Sunlight requirements
          example: "Bright, indirect light. Avoid direct sun."
        watering:
          type: string
          description: Watering instructions
          example: "Water when soil is completely dry (every 2-3 weeks)."
        soil:
          type: string
          description: Soil recommendations
          example: "Well-draining cactus or succulent mix."
        notes:
          type: string
          description: Additional care notes
          example: "Hardy and easy to care for. Great for beginners."
...
    HistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        genus:
          type: string
        species:
          type: string
        confidence:
          type: number
          format: float
        image_path:
          type: string
          description: Image filename (not full path)
        created_at:
          type: string
          format: date-time
    HistoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryItem'
        total:
          type: integer
          description: Total number of identifications
        limit:
          type: integer
          description: Number of items per page
        offset:
          type: integer
          description: Current offset
    HistoryDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        genus:
          type: string
        species:
          type: string
        confidence:
          type: number
          format: float
        image_path:
          type: string
        care_guide:
          $ref: '#/components/schemas/CareInstructions'
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid file type"